+++ packages/bpk-component-input/src/BpkClearButton.js
@@ -1,7 +1,7 @@
 /*
  * Backpack - Skyscanner's Design System
  *
- * Copyright 2016-2019 Skyscanner Ltd
+ * Copyright 2016-2020 Skyscanner Ltd
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -52,7 +52,10 @@ const BpkClearButton = (props: Props) => {
       className={classNames.join(' ')}
       {...rest}
     >
-      <ClearButtonIcon className={getClassName('bpk-clear-button__icon')} />
+      <ClearButtonIcon
+        focusable="false" // prevents focus on IE11
+        className={getClassName('bpk-clear-button__icon')}
+      />
     </button>
   );
 };