+++ packages/bpk-react-utils/src/isRTL-test.js
@@ -1,7 +1,7 @@
 /*
  * Backpack - Skyscanner's Design System
  *
- * Copyright 2016-2019 Skyscanner Ltd
+ * Copyright 2016-2020 Skyscanner Ltd
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,6 +15,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
+/* @flow strict */
 
 import isRTL from './isRTL';
 
@@ -36,12 +38,20 @@ describe('isRTL document defined', () => {
     writable: false,
   });
   it('should return true when documentElement direction attribute is `ltr`', () => {
-    document.documentElement.getAttribute.mockReturnValue('ltr');
+    // eslint-disable-next-line prefer-destructuring
+    const documentElement: ?HTMLElement = document.documentElement;
+    if (documentElement) {
+      documentElement.getAttribute.mockReturnValue('ltr');
+    }
     expect(isRTL()).toBe(false);
   });
 
   it('should return true when documentElement direction attribute is `rtl`', () => {
-    document.documentElement.getAttribute.mockReturnValue('rtl');
+    // eslint-disable-next-line prefer-destructuring
+    const documentElement: ?HTMLElement = document.documentElement;
+    if (documentElement) {
+      documentElement.getAttribute.mockReturnValue('rtl');
+    }
     expect(isRTL()).toBe(true);
   });
 });